+++ src/feed_the_troll/__init__.py
@@ -19,5 +19,6 @@ package.
 # Imports
 ##############################################################################
 
-from . import loaders
+from . import feeders
+from . import servers
 from . import trolls